adding the top 3 customer reqard, however i changed it to give it to the poorest customers instead :)

+++ bank/file_management.py
@@ -92,7 +92,7 @@ class FileManagement():
                         row[f"{key}"] = int(value)
                         continue
                     if self.is_number(value):
-                        row[f"{key}"] = int(value)
+                        row[f"{key}"] = float(value)
